+++ khoros/core.py
@@ -3665,6 +3665,10 @@ class Khoros(object):
             .. versionadded:: 3.5.0
 
             :param fields: One or more fields from the ``Users`` object to return (``login`` field by default)
+
+                           .. seealso:: The fields that can be used are found in the
+                                        `Khoros developer documentation <https://bit.ly/3LQLyW5>`_.
+
             :type fields: str, tuple, list, set
             :param role_id: The identifier for the role in ``node_type:node_id:role_name`` format
             :type role_id: str, None